+++ app/src/main/java/com/ktx/dormitory/student/payment/presentation/PaymentContract.kt
@@ -4,6 +4,7 @@ import android.os.Parcelable
 import com.ktx.dormitory.student.payment.domain.model.Bill
 import com.ktx.dormitory.student.payment.domain.model.PaymentResult
 import kotlinx.parcelize.Parcelize
+import java.math.BigDecimal
 
 sealed class PaymentUiState : Parcelable {
     @Parcelize
@@ -12,7 +13,7 @@ sealed class PaymentUiState : Parcelable {
     @Parcelize
     data class Success(
         val bills: List<Bill>,
-        val totalUnpaid: Double,
+        val totalUnpaid: BigDecimal,
         val isProcessing: Boolean = false,
         val smartQR: PaymentResult? = null // Thông tin QR thông minh nếu đang mở
     ) : PaymentUiState()
